backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.10
redis==5.0.3
celery==5.3.6
django-celery-beat==2.6.0
channels==4.1.0
channels-redis==4.2.0
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3
Pillow==11.1.0
python-decouple==3.8
sentry-sdk==1.45.0
django-ratelimit==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
